Move Captchacheck tests to Jest
