+++ node/risk-app/test/mock.js
@@ -70,26 +70,6 @@ const mockCaptchaCheck = (postcode) => {
   }
 }
 
-const createMockYar = () => {
-  class YarMock {
-    constructor () {
-      this._store = { }
-    }
-
-    get (key) {
-      return this._store[key]
-    }
-
-    set (key, value) {
-      this._store[key] = value
-    }
-  }
-
-  const yar = new YarMock()
-
-  return yar
-}
-
 const mockCaptchaResponse = (responseType, errorType) => {
   if (responseType === false && errorType === 'solution timeout') {
     return { success: false, errors: ['solution_timeout_or_duplicate'] }
@@ -115,4 +95,4 @@ const createAddressStub = (service, newAddress) => {
   return mock.replace(service, 'find', mock.makePromise(null, addressToReplace))
 }
 
-module.exports = { mock, mockOptions, mockSearchOptions, mockCaptchaResponse, createMockYar, mockCaptchaCheck, createWarningStub, createAddressStub }
+module.exports = { mock, mockOptions, mockSearchOptions, mockCaptchaResponse, mockCaptchaCheck, createWarningStub, createAddressStub }
